+++ distributionlist_contact.php
@@ -109,6 +109,7 @@ if($action === 'confirm_add_filter' && $confirm === 'yes') {
 	unset($TParamURL['pageplusone']);
 	unset($TParamURL['limit']);
 }
+
 $TParamURL_HTTP_build_query = http_build_query($TParamURL);
 
 if($action === 'add_filter') {
@@ -165,6 +166,13 @@ if($action === 'add_filter') {
 		$TParamURL_HTTP_build_query = $f->url_params;
 
 	} else setEventMessage($langs->trans('DistributionListNeedToSelectFilterForSelection'), 'warnings');
+}
+//maintient le filtre si on change le nombre de contacts à afficher
+elseif(!empty($filter_id) && GETPOSTISSET('limit')){
+	$f = new DistributionListSocpeopleFilter($db);
+	$res = $f->fetch($filter_id);
+	if($res > 0) $TParamURL_HTTP_build_query = $f->url_params . $TParamURL_HTTP_build_query;
+	else setEventMessage($langs->trans('DistributionListFilterError'), 'errors');
 }
 
 if($action === 'add_all_filtered_contacts') {